move macros to op_regist.h
