+++ chainerx_cc/chainerx/native/native_device/arithmetic.cc
@@ -8,7 +8,6 @@
 #include "chainerx/dtype.h"
 #include "chainerx/float16.h"
 #include "chainerx/native/elementwise.h"
-#include "chainerx/native/native_device/std_ops.h"
 #include "chainerx/native/op_regist.h"
 #include "chainerx/routines/math.h"
 #include "chainerx/scalar.h"
